+++ packages/graph-explorer/src/connector/sparql/templates/classesWithCountsTemplates.ts
@@ -1,8 +1,8 @@
-import dedent from "dedent";
+import { query } from "@/utils";
 
 // It returns a list of classes with the number of instances for each class
 export default function classesWithCountsTemplates() {
-  return dedent`
+  return query`
     # Fetch a list of classes with the number of instances for each class
     SELECT ?class  (COUNT(?start) AS ?instancesCount) {
       ?start a ?class