+++ packages/graph-explorer/src/utils/index.ts
@@ -9,5 +9,6 @@ export { DEFAULT_SERVICE_TYPE } from "./constants";
 export { default as escapeString } from "./escapeString";
 export { default as batchPromisesSerially } from "./batchPromisesSerially";
 export { default as logger } from "./logger";
+export * from "./sanitizeQuery";
 export * from "./set";
 export * from "./env";