+++ Changelog.md
@@ -4,6 +4,8 @@
 
 - **Fixed** Docker image containing more files than necessary.
   ([#613](https://github.com/aws/graph-explorer/pull/613))
+- **Improved** query generation by removing empty lines
+  ([#608](https://github.com/aws/graph-explorer/pull/608))
 
 ## Release 1.10.1